--- .gitignore
@@ -1,3 +1,0 @@
-.env
-node_modules
-route.rest
