+++ modules/subscriber.js
@@ -1,0 +1,19 @@
+const mongoose = require('mongoose')
+
+const subscriberSchema = new mongooose.schema({
+    name: {
+        type: String,
+        required: true
+    },
+    subscriberToChannel: {
+        type: String,
+        required: true
+    },
+    subscribeDate: {
+        type: Date,
+        required: true,
+        default: Date.now
+    }
+})
+
+module.exports = mongoose.model('Subscriber', subscriberScheme)
