+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "current-project",
+  "name": "package",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -8,12 +8,8 @@
   },
   "author": "",
   "license": "ISC",
-  "dependencies": {
-    "express": "^4.16.4",
-    "mongoose": "^5.5.7"
-  },
   "devDependencies": {
-    "dotenv": "^8.6.0",
-    "nodemon": "^1.19.4"
+    "dotenv": "^16.0.3",
+    "nodemon": "^2.0.20"
   }
 }
