+++ server.js
@@ -1,17 +1,16 @@
-require('dotenv').config()
-
 const express = require('express')
 const app = express()
 const mongoose = require('mongoose')
 
-mongoose.connect('mongodb+srv://Pegasusx10:pegasus123@cluster0.749smlf.mongodb.net/student-database?retryWrites=true&w=majority', { useUnifiedTopology: true, useNewUrlParser: true })
+mongoose.connect(process.env.DATABASE_URL, { useNewUrlParser: true})
 const db = mongoose.connection
 db.on('error', (error) => console.error(error))
-// db.once('open', () => console.log('Connected to Database'))
+db.once('open', ()=> console.log('connected to Database'))
 
 app.use(express.json())
 
-const studentsRouter = require('./routes/students')
-app.use('/students', studentsRouter)
 
-app.listen(5000, () => console.log('Server Started'))
+const subscribersRouter = require('-/routes/subscribers')
+app.use('/subcribers', subscriberRouter)
+
+app.listen(5000, () => console.log('server started'))
